fix(pipeline): address batch3 review minor findings
- runstore: surface lock-file close errors and remove the lock on partial
  writes instead of swallowing Close in a deferred call
- parse_blocks: only treat a REPLACE section as well-formed when it is
  closed by a ">>>>>>>" marker; a "=======" without one now returns the
  original content instead of capturing trailing garbage
- controller: compare the short SHA form of HEAD against the ledger's short
  commit when deciding whether to prune a lost-commit task, fixing a
  full-vs-short mismatch that could wrongly prune the branch tip
- controller_test: elide the 5s dispatch-retry backoff in
  TestCheckpointLogsError (down to ~10ms) and add TestRunChainsPriorOnTokens,
  TestRecoveryKeepsTipCommitWithShortSHA, and a malformed-REPLACE guard test

+++ internal/pipeline/controller.go
@@ -1134,8 +1134,10 @@ func (c *Controller) Run(ctx context.Context) error {
 			continue // commit exists; keep the task done
 		}
 		// LogOneline errored — the commit may be missing. Only prune if it
-		// is also not the current branch tip.
-		if h, err := c.Git.RevParse(c.workDir(), "HEAD"); err == nil && h == head {
+		// is also not the current branch tip. The ledger stores the short
+		// SHA form, so compare against the short form of HEAD to avoid a
+		// full-vs-short mismatch that would never match.
+		if h, err := c.Git.RevParse(c.workDir(), "HEAD"); err == nil && short(h) == head {
 			continue // head is the branch tip; keep the task done
 		}
 		delete(done, taskN)
